隐藏团队收益

+++ Assets/lb_event.cs
@@ -7,10 +7,10 @@ public class lb_event : MonoBehaviour {
 
 	public void Getsl(string value)
     {
-        Debug.Log(int.Parse(value));
-        if (int.Parse(value) > 0)
+        Debug.Log(float.Parse(value));
+        if (float.Parse(value) > 0)
             this.GetComponent<Text>().color = new Color(0.13f,0.56f,0.13f);
-        if (int.Parse(value) < 0)
+        if (float.Parse(value) < 0)
             this.GetComponent<Text>().color = new Color(0.56f, 0.13f, 0.13f);
 
     }
